refactor: Move run function to lib
