+++ src/main.rs
@@ -2,15 +2,7 @@ use std::str::FromStr;
 
 use anyhow::{Context, Result};
 use clap::{Arg, Command};
-use settings::Settings;
 use time::macros::format_description;
-
-mod database;
-mod model;
-mod server;
-mod settings;
-#[cfg(test)]
-mod tests;
 
 fn setup_logging(level: &str) -> Result<()> {
     let level = log::LevelFilter::from_str(level).unwrap();
@@ -63,34 +55,6 @@ async fn main() -> Result<()> {
 
     setup_logging(opts.get_one::<String>("log_level").unwrap())?;
 
-    let settings =
-        Settings::load(opts.get_one::<String>("config").unwrap()).context("can't load config.")?;
-
-    let (tx, rx) = tokio::sync::mpsc::channel::<model::Report>(64);
-
-    let server = {
-        let settings = settings.server;
-        tokio::spawn(async move {
-            let tx = tx.clone();
-            server::run_server(settings, tx).await.unwrap();
-        })
-    };
-
-    {
-        let settings = settings.database.clone();
-        tokio::spawn(async move {
-            database::aggregate_loop(&settings).await;
-        });
-    }
-
-    {
-        let settings = settings.database;
-        tokio::spawn(async move {
-            database::insert_reports_loop(&settings, rx).await;
-        });
-    }
-
-    server.await?;
-
+    barad_dur::run(opts).await?;
     Ok(())
 }
